+++ lib/mvvm_service.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'components/hook_ticker_provider.dart';
 export 'components/service_mode.dart';
 export 'components/service_subscription.dart';
 export 'components/service.dart';
@@ -7,6 +8,7 @@ export 'components/state_id.dart';
 export 'components/state_mode.dart';
 export 'components/state_subscription.dart';
 
+export 'extensions/hooks_build_extension.dart';
 export 'extensions/service_build_context.dart';
 export 'extensions/service_when.dart';